backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.0.3
openai==1.3.5
anthropic==0.7.1
httpx==0.25.1
praw==7.7.1
feedparser==6.0.10
schedule==1.2.0
groq==0.4.1
huggingface-hub>=1.1.0
